Use fully qualified image name
In OpenShift, if you use a pull secret for Docker Hub, it will only be used if the fully qualified image name is used. The common reason for a pull secret is to deal with the Docker Hub rate limit.

Otherwise I have to overwrite with `baseimage` to work around the rate limit.