Add environment variables configuring setting dashboard namespaces
Currently it's only possible to configure dashboard namespaces through
command line args. This makes it difficult to fine-tune the operator
when installing it declaratively through an OLM subscription.

This commit adds the DASHBOARD_NAMESPACES and DASHBOARD_NAMESPACES_ALL
environment variables which serve as fallback for the namespaces and
scan-all flags.