bugfix - datasource controller was dropping secureJsonData
This addresses a bug introduced in the changeover to using the openapi
Grafana library where the secureJsonData field was being dropped.

To address this it's necessary to use the `UpdateDataSourceCommand`
struct to ensure we're also not dropping the version if it's present.

This is a bit counter-intuitive to use the `Update` struct here and
perhaps worth looking at another solution at some point.