fix: improve nested folder handling
This fixes some issues around the move logic. The underlying issue was the way
in which folder existence was checked. As the search folder endpoint does _not_
return nested folders, moving between folders was not possible. This fix prefers
matching by UIDs first.

Note that the introduction of nested folders implies that they will only work
with folders created by the operator as names no longer need to be unique and
the only way to keep track is to use the UID.