fix(helm): allow setting pod labels
Since #1538 landed, it's not possible to set additional pod labels. The
only way of setting pod labels would be to change
`grafana-operator.selectorLabels`, but doing that breaks current
deployments as selector labels are immutable.

`grafana-operator.labels` includes `grafana-operator.selectorLabels`, so
it's safe to include them instead in the pod labels, and doing so fixes
the current issue without causing side-effects.

fixes #1546 #1642