disallow fetching library panels from grafana.com
much of the content resolving/fetching logic is shared b/w dashboards
and library panels, but there is one awkward thing where library panels
are not currently (to my knowledge) available via grafana.com. rather
than extract just that fetch logic out and only apply it to dashboards,
i've opted to simply disable them as a source. an error will be issued
if a user tries to configure this field. it is not the most graceful
way to handle this, as it's not communicated via the CR validation, but
given the effort required to bend the code to do this "right", and
given the likelihood of somebody trying to configure a library panel
via a grafana.com url (which does not exist right now), we are choosing
the easy button.