chore(controllers): extract dashboard model handling logic to be later used for library panels (#1845)
* extract dashboard model handling logic

the motivation for this is to enable management of Library Panels within
the operator. Library Panels share many of the same ergonomics around
Dashboards: they are ultimately a JSON model understood by grafana, adhering
to the same uid/id semantics, support dashboard references, and can be
theoretically sourced in a variety of means: configMap, URL, raw JSON, etc.

this extracts all of the shared logic into a separate "content" package, and
introduces a ContentResolver mechanism that performs the same logic that used
to be implemented directly by the GrafanaDashboard.

to support resolving against different types of resources, this also introduces
a GrafanaContentResource interface, and updates GrafanaDashboard to implement
that interface. many of the common spec and status fields were likewise extracted
into common structures; the interface's main purpose is to provide uniform access
to those structures.

this should not change any core business logic; tests have been moved around to
their new logical location.

to resolve issues of circular imports, the "fetchers" package was moved to be
a subpackage of "content", and the gzip logic was moved into a separate "cache"
subpackage.

* do not requeue after unrecoverable error