enable separate metrics for content resources
to enable dashboards and library panels to track their fetch operations
separately, split up the metrics by resource. this is accomplished by
adding a new interface method to the GrafanaContentResource, and the
returned struct holds references to the relevant metrics, if any.

this also adds an additional guard to the http round tripper so that if
no metric is defined, we don't panic.