move metrics out of api module
since #1856, the api module is separated. this poses problems
with the approach to have the CRs expose which metrics correspond
to their types via the GrafanaContentMetrics() function, because
we create an import loop.

recognizing that we're measuring essentially the same thing b/w
library panels and dashboards, this updates the metrics to indicate
this more explicitly.

the following changes are made to metrics:

  `grafana_operator_dashboard_requests`:
    * renamed to `grafana_operator_content_requests`, but will always have a label
      `kind="GrafanaDashboard"`.
    * renamed `dashboard` label to `resource`

  `grafana_operator_dashboard_revision_requests`:
    * renamed to `grafana_operator_revision_requests`, but will always have a label
      `kind="GrafanaDashboard"`.
    * renamed `dashboard` label to `resource`