add backwards-compatibility for metrics and simplify reconcile
to make the dashboard url request metric backwards compatible, this
updates the signature of the roundtripper to be a bit more general,
so we can pass a function that is called at the end of the response
rather than explicitly increment a metric. this allows us to increment
multiple metrics in the case of the dashboard url request, one for the
old and one for the new.