feat: add GrafanaLibraryPanel CRD and reconciler (#1858)
* add grafanalibrarypanel CRD types

this adds a new GrafanaLibraryPanel CRD, updates all the auto-generated
assets derived from the CRD type, and updates existing RBAC and Helm
definitions to reference the new type.

Co-authored-by: Matt Delaney <matt.delaney@getcruise.com>

* track library panels in grafana CR

this registers the new CRD with the Grafana CRD so that we can track
which library panels have been provisioned with the operator, similar
to how other resources are tracked.

Co-authored-by: Matt Delaney <matt.delaney@getcruise.com>

* add reconciler implementation

Co-authored-by: Matt Delaney <matt.delaney@getcruise.com>

* disallow fetching library panels from grafana.com

much of the content resolving/fetching logic is shared b/w dashboards
and library panels, but there is one awkward thing where library panels
are not currently (to my knowledge) available via grafana.com. rather
than extract just that fetch logic out and only apply it to dashboards,
i've opted to simply disable them as a source. an error will be issued
if a user tries to configure this field. it is not the most graceful
way to handle this, as it's not communicated via the CR validation, but
given the effort required to bend the code to do this "right", and
given the likelihood of somebody trying to configure a library panel
via a grafana.com url (which does not exist right now), we are choosing
the easy button.

* remove dashboard name from generic functions

* add chainsaw tests

* mark proposal as decided

* enable separate metrics for content resources

to enable dashboards and library panels to track their fetch operations
separately, split up the metrics by resource. this is accomplished by
adding a new interface method to the GrafanaContentResource, and the
returned struct holds references to the relevant metrics, if any.

this also adds an additional guard to the http round tripper so that if
no metric is defined, we don't panic.

* remove lazy folder creation and simplify logic

* move metrics out of api module

since #1856, the api module is separated. this poses problems
with the approach to have the CRs expose which metrics correspond
to their types via the GrafanaContentMetrics() function, because
we create an import loop.

recognizing that we're measuring essentially the same thing b/w
library panels and dashboards, this updates the metrics to indicate
this more explicitly.

the following changes are made to metrics:

  `grafana_operator_dashboard_requests`:
    * renamed to `grafana_operator_content_requests`, but will always have a label
      `kind="GrafanaDashboard"`.
    * renamed `dashboard` label to `resource`

  `grafana_operator_dashboard_revision_requests`:
    * renamed to `grafana_operator_revision_requests`, but will always have a label
      `kind="GrafanaDashboard"`.
    * renamed `dashboard` label to `resource`

* add backwards-compatibility for metrics and simplify reconcile

to make the dashboard url request metric backwards compatible, this
updates the signature of the roundtripper to be a bit more general,
so we can pass a function that is called at the end of the response
rather than explicitly increment a metric. this allows us to increment
multiple metrics in the case of the dashboard url request, one for the
old and one for the new.

* refactor: pass metrics to round_tripper directly

* refactor: remove intermediate structs to further simplify logic

* register invalidspec condition on validation errs

* Update api/v1beta1/grafanalibrarypanel_types.go

Co-authored-by: Steffen Baarsgaard <steff.bpoulsen@gmail.com>

* regen

---------

Co-authored-by: Matt Delaney <matt.delaney@getcruise.com>
Co-authored-by: Dominik Süß <dominik@suess.wtf>
Co-authored-by: Steffen Baarsgaard <steff.bpoulsen@gmail.com>