chore: add docs for librarypanel CR
this adds documentation for the library panel resource type. since most
of its features are shared with dashboards when it comes to provisioning,
we just link to the dashboard documentation instead of calling out each
feature individually.

exporting panels from grafana is unfortunately rather tedious. i've
included instructions on how to reliably do this, and some examples. i
hope that in the future the story around library panel sharing improves;
for now, this gets the job done.