feat: allow customizing admin url when using k8s service (#1874)
* feat: allow customizing admin domain when using k8s service

this is a no-op by default, but the fqdn can help when other network policy agents
are in the mix, which may have rules set up for svc.cluster.local to
enable traffic passing. for example, in an environment with istio and opa in
the mix, there might be policies and/or assumptions about what form the service
address takes. we already saw this with openshift's certificate signing; this just
opens a door to other exceptions in the future.

* always include .svc in the host address

* revise docs and rename to clusterDomain

clusterDomain is how this setting is generally referred to in kubernetes;
aligning on that terminology for consistency.

---------

Co-authored-by: Igor Beliakov <46579601+weisdd@users.noreply.github.com>