feat: add new flag max-concurrent-reconciles (#1969)
* feat: add new flag max-concurrent-reconciles

This is to allow control for users to set proper value for MaxConcurrentReconciles for the certain controllers.

Only applies to high volume CR reconcilers(datasource, dashboards, folders) given the rest do not seem to be high volume.

It is necessary to speed up reconcile time for certain conditions. Whenever thousands of CRs being reconciled either during start up or re-deploy of the operator or hot-sync, it takes quite some time.

Signed-off-by: emreya <e.yazici1990@gmail.com>

* chore: set max-concurrent-reconciles on kustomize & helm

To let users beware of the flag.

Signed-off-by: emreya <e.yazici1990@gmail.com>

* chore: comment out setting new flag

will be moved to doc eventually.

Signed-off-by: emreya <e.yazici1990@gmail.com>

* chore: set maxConcurrentReconciles at manager level

As the vision is to be consistent for all contrllers and expect to have many more grafana CRs even, moving this option to global level makes sense.

Controller specific opts will be added in the future if the need arises.

Signed-off-by: emreya <e.yazici1990@gmail.com>

* chore: correct import order

Signed-off-by: emreya <e.yazici1990@gmail.com>

* chore: fix maxConcurrentReconciles in Helm, yaml formatting, run helm-docs

---------

Signed-off-by: emreya <e.yazici1990@gmail.com>
Co-authored-by: Steffen Baarsgaard <steff.bpoulsen@gmail.com>