feat: add new flag default-resync-period (#2222)
* feat: add new flag default-resync-period

This is to allow control for users to set value for resync-period in case necessary.

Any explicit setting on CRs will take precedence over this setting.

* chore: remove Resync default from CRD

This is to be able to configure global interval through flags on the operator.

Default for the flag `default-resync-period` set to 10m for unchanged behaviour.

* refactor: avoid unnecessary alias on import

* refactor: make Config as ptr

It is shared across whole controllers. Thus, semantically ptr is better.

* chore: bring lost changes during rebase