feat: use kong to parse flags (#2289)
* feat: use kong to parse flags

This change introduces [kong](https://github.com/alecthomas/kong) to parse
command line flags. This also allows us to use environment variables in
configuring the operator

This is done in preparation to add a feature flag system.

* chore(lints): Address Whitespace, error and security lint for int8 conversion

* chore: Format struct tags and use kong enum input validation

I would have liked to do it for ZapLogLevel as well, but then that would break parsing ints

---------

Co-authored-by: Steffen Baarsgaard <steff.bpoulsen@gmail.com>